feat: full auth system + OTP + JWT + PIN + SMS fix

+++ src/main/java/com/kuza/kuzasokoni/domain/product/dtos/command/TenureCreateCommand.java
@@ -23,7 +23,7 @@ public class TenureCreateCommand {
         private Integer phase3Days;
         private Integer phase4Days;
 
-        private Integer penaltyCap;
+        private Double penaltyCap;
         private Integer gracePeriodDays;
 
          private BigDecimal interest;
