+++ src/main/java/com/kuza/kuzasokoni/domain/product/dtos/command/TenureUpdateCommand.java
@@ -23,7 +23,7 @@ import java.math.BigDecimal;
         private Integer phase3Days;
         private Integer phase4Days;
 
-        private Integer penaltyCap;
+        private Double penaltyCap;
         private Integer gracePeriodDays;
 
         private BigDecimal interest;
